move the node test to the binary_quantized_tests
