+++ src/tests/mod.rs
@@ -10,7 +10,6 @@ use crate::roaring::RoaringBitmapCodec;
 use crate::{Database, Distance, MetadataCodec, NodeCodec, NodeMode, Reader};
 
 mod binary_quantized;
-mod node;
 mod reader;
 mod writer;
